requirements-dev.txt
mypy==1.13.0
black
flake8==7.1.1
flake8-django
pylint==3.3.1
django-linter
django-stubs
pytest==8.3.3
pytest-django
make
astroid==3.3.5
